OSIS-4985-bis : Modifications après review
