+++ education_group/views/group/utilization_read.py
@@ -25,7 +25,7 @@
 ##############################################################################
 from education_group.views.group.common_read import Tab, GroupRead
 from program_management.ddd.domain.node import NodeIdentity
-from program_management.views.element_utilization import get_utilization_rows
+from program_management.ddd.service.read.get_utilization_rows import get_utilizations
 
 
 class GroupReadUtilization(GroupRead):
@@ -35,5 +35,5 @@ class GroupReadUtilization(GroupRead):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         node = self.get_object()
-        context['utilization_rows'] = get_utilization_rows(NodeIdentity(code=node.code, year=node.year))
+        context['utilization_rows'] = get_utilizations(NodeIdentity(code=node.code, year=node.year))
         return context
