+++ program_management/ddd/domain/program_tree.py
@@ -24,7 +24,7 @@
 #
 ##############################################################################
 from collections import Counter
-from typing import List, Set, Optional
+from typing import List, Set, Optional, Dict
 
 import attr
 
@@ -575,3 +575,20 @@ def _copy(root: 'Node', ignore_children_from: Set[EducationGroupTypesEnum] = Non
         new_children.append(new_link)
     new_node.children = new_children
     return new_node
+
+
+def get_nearest_parents(parents: List['Node']) -> Optional[Dict[str, 'Node']]:
+    parent_direct_node = None
+    root_node = None
+    for node in parents:
+        if (node.is_minor_or_deepening()) or (node.is_training() and node.is_finality()):
+            parent_direct_node = node
+        if node.is_training() or node.is_mini_training():
+            root_node = node
+
+    if parent_direct_node is None and root_node is None:
+        return None
+    return {
+        'parent_direct_node': parent_direct_node,
+        'root_node': root_node
+    }
